refactor: extract ViewMission and ViewRuntime in another file
